requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
attrs==23.1.0
braceexpand==0.1.7
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
decorator==4.4.2
Farama-Notifications==0.0.4
faster-fifo==1.4.5
filelock==3.12.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==0.4.6
grpcio==1.54.2
gym==0.23.0
gym-notices==0.0.8
gymnasium==0.28.1
-e git+https://github.com/facebookresearch/habitat-lab.git@332f4c48033b3d985b936a1abe1b3452ce556c77#egg=habitat_baselines&subdirectory=habitat-baselines
-e git+https://github.com/facebookresearch/habitat-lab.git@332f4c48033b3d985b936a1abe1b3452ce556c77#egg=habitat_lab&subdirectory=habitat-lab
habitat-sim @ file:///Users/work/habitat-sim
hydra-core==1.3.2
idna==3.4
ifcfg==0.24
imageio==2.28.1
imageio-ffmpeg==0.4.8
iniconfig==2.0.0
jax-jumpy==1.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
llvmlite==0.40.0
lmdb==1.4.1
magnum @ file:///Users/work/habitat-sim/build/deps/magnum-bindings/src/python
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.0
numpy==1.24.3
numpy-quaternion==2022.4.3
oauthlib==3.2.2
objectio==0.2.29
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
proglog==0.1.10
protobuf==3.20.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.0.4
pygame==2.4.0
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sympy==1.12
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==2.0.1
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.5.0
urllib3==1.26.15
webdataset==0.1.40
Werkzeug==2.3.4
